Introduce FlinkSource
